Add and hook up volumes list component
