+++ src/linodes/volumes/layouts/IndexPage.js
@@ -2,28 +2,12 @@ import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
 
 import { PrimaryButton } from 'linode-components/buttons';
-import { Input } from 'linode-components/forms';
-import { List } from 'linode-components/lists';
-import { Table } from 'linode-components/tables';
-import { MassEditControl } from 'linode-components/lists/controls';
-import { ListHeader } from 'linode-components/lists/headers';
-import { ListBody } from 'linode-components/lists/bodies';
-import { DeleteModalBody } from 'linode-components/modals';
-import {
-  ButtonCell,
-  CheckboxCell,
-  LabelCell,
-} from 'linode-components/tables/cells';
-import { RegionCell } from '~/components/tables/cells';
 
 import { setAnalytics, setSource, setTitle } from '~/actions';
-import { showModal, hideModal } from '~/actions/modal';
-import { default as toggleSelected } from '~/actions/select';
 import { volumes, linodes } from '~/api';
-import { transform } from '~/api/util';
 import CreateHelper from '~/components/CreateHelper';
 
-import { AddVolume } from '../components';
+import { AddEditVolume, VolumesList } from '../components';
 
 
 const OBJECT_TYPE = 'volumes';
@@ -34,12 +18,6 @@ export class IndexPage extends Component {
     dispatch(linodes.all());
   }
 
-  constructor(props) {
-    super(props);
-
-    this.state = { filter: '' };
-  }
-
   async componentDidMount() {
     const { dispatch } = this.props;
     dispatch(setSource(__filename));
@@ -47,107 +25,8 @@ export class IndexPage extends Component {
     dispatch(setTitle('Volumes'));
   }
 
-  deleteVolumes = (volumesToDelete) => {
-    const { dispatch } = this.props;
-    const volumesArr = Array.isArray(volumesToDelete) ? volumesToDelete : [volumesToDelete];
-
-    const selectedVolumes = volumesArr.map(l => l.label);
-
-    dispatch(showModal('Delete Volume(s)', (
-      <DeleteModalBody
-        onSubmit={async () => {
-          const ids = volumesArr.map(function (volume) { return volume.id; });
-
-          await Promise.all(ids.map(id => dispatch(volumes.delete(id))));
-          dispatch(toggleSelected(OBJECT_TYPE, ids));
-          dispatch(hideModal());
-        }}
-        items={selectedVolumes}
-        typeOfItem="Volumes"
-        onCancel={() => dispatch(hideModal())}
-      />
-    )));
-  }
-
-  renderVolumes(volumes) {
-    const { dispatch, selectedMap } = this.props;
-    const { filter } = this.state;
-
-    const { sorted } = transform(volumes, {
-      filterBy: filter,
-      sortBy: v => v.label.toLowerCase(),
-    });
-
-    return (
-      <List>
-        <ListHeader className="Menu">
-          <div className="Menu-item">
-            <MassEditControl
-              data={sorted}
-              dispatch={dispatch}
-              massEditGroups={[{ elements: [
-                { name: 'Delete', action: this.deleteVolumes },
-              ] }]}
-              selectedMap={selectedMap}
-              objectType={OBJECT_TYPE}
-              toggleSelected={toggleSelected}
-            />
-          </div>
-          <div className="Menu-item">
-            <Input
-              placeholder="Filter..."
-              onChange={({ target: { value } }) => this.setState({ filter: value })}
-              value={this.state.filter}
-            />
-          </div>
-        </ListHeader>
-        <ListBody>
-          <Table
-            columns={[
-              { cellComponent: CheckboxCell, headerClassName: 'CheckboxColumn' },
-              {
-                cellComponent: LabelCell,
-                headerClassName: 'LabelColumn',
-                dataKey: 'label',
-                titleKey: 'label',
-                tooltipEnabled: true,
-              },
-              { dataFn: (volume) => {
-                const { size } = volume;
-                return `${size} GiB`;
-              } },
-              {
-                cellComponent: RegionCell,
-                headerClassName: 'RegionColumn',
-              },
-              { dataFn: (volume) => {
-                const { linode_id: linodeId } = volume;
-                if (!linodeId) {
-                  return 'Unattached';
-                }
-                return `Attached to ${linodeId}`;
-              } },
-              {
-                cellComponent: ButtonCell,
-                headerClassName: 'ButtonColumn',
-                text: 'Delete',
-                onClick: (volume) => { this.deleteVolumes(volume); },
-              },
-            ]}
-            data={sorted}
-            selectedMap={selectedMap}
-            disableHeader
-            onToggleSelect={(record) => {
-              dispatch(toggleSelected(OBJECT_TYPE, record.id));
-            }}
-          />
-        </ListBody>
-      </List>
-    );
-  }
-
   render() {
-    const { dispatch, linodes } = this.props;
+    const { dispatch, linodes, volumes, selectedMap } = this.props;
 
     return (
       <div className="PrimaryPage container">
@@ -156,15 +35,20 @@ export class IndexPage extends Component {
             <h1 className="float-sm-left">Volumes</h1>
             <PrimaryButton
               className="float-sm-right"
-              onClick={() => AddVolume.trigger(dispatch, linodes)}
+              onClick={() => AddEditVolume.trigger(dispatch, linodes)}
             >
               Add a Volume
             </PrimaryButton>
           </div>
         </header>
         <div className="PrimaryPage-body">
-          {Object.keys(this.props.volumes.volumes).length ?
-            this.renderVolumes(this.props.volumes.volumes) :
+          {Object.keys(volumes.volumes).length ?
+           <VolumesList
+             objectType={OBJECT_TYPE}
+             volumes={volumes.volumes}
+             selectedMap={selectedMap}
+             dispatch={dispatch}
+           /> :
             <CreateHelper
               label="Volumes"
               onClick={() => {}}
